conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altgraph=0.17=pyhd8ed1ab_0
ca-certificates=2020.10.14=0
certifi=2021.5.30=py39haa95532_0
click=8.0.1=py39hcbf5309_0
colorama=0.4.4=pyh9f0ad1d_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py39hcbf5309_3
icu=58.2=ha925a31_3
importlib-metadata=3.10.0=py39haa95532_0
joblib=0.17.0=py_0
jpeg=9b=hb83a4c4_2
libpng=1.6.37=h1d00b33_2
libtiff=4.2.0=hd0e1b90_0
lz4-c=1.9.3=h2bbff1b_1
macholib=1.14=pyh9f0ad1d_1
olefile=0.46=py_0
openssl=1.1.1k=h2bbff1b_0
pefile=2019.4.18=py_0
piexif=1.1.3=py_2
pillow=8.3.1=py39h4fa10fc_0
pip=21.2.2=py39haa95532_0
pycryptodome=3.10.1=py39hb82d6ee_0
pyinstaller=3.6=py39h8cc25b3_6
pyqt=5.9.2=py39hd77b12b_6
python=3.9.6=h7840368_1_cpython
python_abi=3.9=2_cp39
pywin32=228=py39hbaba5e8_1
pywin32-ctypes=0.2.0=py39hcbf5309_1003
qt=5.9.7=vc14h73c81de_0
setuptools=52.0.0=py39haa95532_0
shortuuid=1.0.1=py39hcbf5309_4
sip=4.19.13=py39hd77b12b_0
sqlite=3.36.0=h8ffe710_0
tk=8.6.10=he774522_0
tzdata=2021a=he74cb21_1
ucrt=10.0.20348.0=h57928b3_0
vc=14.2=hb210afc_5
vs2015_runtime=14.29.30037=h902a5da_5
wheel=0.36.2=pyhd3deb0d_0
wincertstore=0.2=py39hcbf5309_1006
xz=5.2.5=h62dcd97_0
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h62dcd97_1010
zstd=1.4.9=h19a0ad4_0
